feat: restructured tables using sqlalchemy
