+++ alembic/env.py
@@ -9,7 +9,7 @@ from sqlalchemy.ext.asyncio import async_engine_from_config
 from alembic import context
 from app.db import Base
 from app.auth.models import User  # New
-from app.songs.tables import Song  # New
+from app.songs.models import Song, City, Tag, SongTag  # New
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
