+++ app/main.py
@@ -3,19 +3,19 @@ from fastapi.middleware.cors import CORSMiddleware
 from fastapi.responses import HTMLResponse
 from fastapi.templating import Jinja2Templates
 
-from .auth.handlers_scopes import router as auth_router
-# from .files.handlers import router as files_router
-# from .live_socket.handlers import router as ws_router
-# from .nosql.handlers import router as nosql_router
-# from .songs.handlers import router as song_router
+from .auth.handlers import router as auth_router
+from .files.handlers import router as files_router
+from .live_socket.handlers import router as ws_router
+from .nosql.handlers import router as nosql_router
+from .songs.handlers import router as song_router
 
 app = FastAPI()
 
-# app.include_router(song_router)
+app.include_router(song_router)
 app.include_router(auth_router)
-# app.include_router(files_router)
-# app.include_router(ws_router)
-# app.include_router(nosql_router)
+app.include_router(files_router)
+app.include_router(ws_router)
+app.include_router(nosql_router)
 
 origins = [
     "http://localhost.tiangolo.com",
